Create an option for enabling/disabling numeric answers (solution to issue #19)
Some spam bots input randomized numbers to the MathSpamProtectorField which some times happen to be correct and let spam come through. This option makes it possible to accept only verbal answers. This commit won't change the default behavior, so in order to disable numeric answers, put this to your yml file:

MathSpamProtectorField:
  allow_numeric_answer: false